Add localStorage property to save local key-values of cart items and refactor production code
