+++ src/App.js
@@ -3,7 +3,7 @@ import "./index.css"
 import { Routes, Route } from "react-router-dom"
 import ShopLayout from "./components/layout/ShopLayout"
 import { Products } from "./components/products/Products.jsx"
-import { Home } from "./pages/Home"
+import { Home } from "./pages/Home/Home"
 import { Error404 } from "./pages/404/Error404"
 import { ProductsDetails } from "./pages/productDetails/ProductsDetails"
 import { Profile } from "./pages/Profile/Profile.jsx"
