+++ src/components/Forms/LoginForm.jsx
@@ -37,7 +37,6 @@ export const LoginForm = () => {
         isClosable: true,
       })
     }
-    console.log(data)
     dispatch(login(data))
   }
