--- src/components/auth/Login.jsx
@@ -1,17 +1,0 @@
-import { Box, Heading, Text } from "@chakra-ui/react"
-
-const Profile = () => {
-  return (
-    <Box>
-      <Heading>hola</Heading>
-      <Text>
-        Lorem ipsum dolor sit, amet consectetur adipisicing elit. Et qui quae
-        nostrum ad provident ut minima magni deleniti. Sapiente quam,
-        dignissimos deserunt delectus labore amet temporibus fugit qui pariatur
-        quos?
-      </Text>
-    </Box>
-  )
-}
-
-export { Profile }
